General - Restructured the project & Optimized Zustand stores & Removed unused Cannon physics engine & Refactored stores
