+++ client/package.json
@@ -9,7 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@react-three/cannon": "^6.5.2",
     "@react-three/drei": "^9.58.5",
     "@react-three/fiber": "^8.12.0",
     "@react-three/postprocessing": "^2.7.1",
