+++ client/src/components/Game/Engine/Engine.context.ts
@@ -1,0 +1,8 @@
+import { createContext } from 'react'
+
+import { Context } from 'react'
+
+type ContextProps = null
+
+export const engineContext: Context<ContextProps> =
+  createContext<ContextProps>(null)
