+++ client/src/components/Game/Engine/Engine.interface.ts
@@ -1,0 +1,3 @@
+export interface EngineProps {
+  children: JSX.Element
+}
