+++ client/src/components/Game/Engine/Environment/Environment.interface.ts
@@ -1,0 +1,5 @@
+import { ThreeEvent } from '@react-three/fiber'
+
+export interface EnvironmentProps {
+  onMoveToPoint: (event: ThreeEvent<MouseEvent>) => void
+}
