+++ client/src/components/Game/Loading/Loading.tsx
