+++ client/src/components/Game/models/Rogue/Rogue.interface.ts
