+++ client/src/components/UserInterface/StatsUI/StatsUI.tsx
@@ -1,44 +1,50 @@
-import { usePlayerUnit } from '@/hooks/usePlayerUnit'
 import { useUnitsStore } from '@/store/units/useUnitsStore'
-
-import _ from 'lodash'
+import { usePlayer } from '@/hooks/usePlayer'
 import { useMemo } from 'react'
+
+import { shallow } from 'zustand/shallow'
+import _ from 'lodash'
+
+import { FindUnit } from '@/store/units/interface'
 
 import styles from './StatsUI.module.scss'
 
 export function StatsUI(): JSX.Element {
-  const { hero } = usePlayerUnit()
-  const findUnit = useUnitsStore().findUnit
+  const { playerHero } = usePlayer()
+
+  const findUnit: FindUnit = useUnitsStore(({ findUnit }) => findUnit, shallow)
 
   const targetUnit = useMemo(() => {
-    if (_.isNil(hero?.target) || hero?.target == null) {
+    if (_.isNil(playerHero?.target) || playerHero?.target == null) {
       return
     }
 
-    return findUnit(hero.target)
-  }, [findUnit, hero?.target])
+    return findUnit(playerHero.target)
+  }, [findUnit, playerHero?.target])
 
-  if (_.isNil(hero)) {
+  if (_.isNil(playerHero)) {
     return <span />
   }
 
   return (
     <div className={styles.container}>
       <p>
-        ⚔ Damage [{hero.attack.type}]: {hero.attack.baseDamage}
+        ⚔ Damage [{playerHero.attack.type}]: {playerHero.attack.baseDamage}
       </p>
 
       <p>
-        🛡 Armor [{hero.defence.type}]: {hero.defence.value}
+        🛡 Armor [{playerHero.defence.type}]: {playerHero.defence.value}
       </p>
 
-      <p>🦵 Agility: {hero.agility}</p>
-      <p>💪 Strength: {hero.strength}</p>
-      <p>📜 Intellect: {hero.intellect}</p>
+      <p>🦵 Agility: {playerHero.agility}</p>
+      <p>💪 Strength: {playerHero.strength}</p>
+      <p>📜 Intellect: {playerHero.intellect}</p>
+      <p>📜 X: {playerHero.position[0]}</p>
+      <p>📜 Z: {playerHero.position[2]}</p>
       <p>
         <b>Target name: {targetUnit?.name ?? 'None'}</b>
       </p>
-      <p>State: {hero.state}</p>
+      <p>State: {playerHero.state}</p>
     </div>
   )
 }
