--- client/src/components/_Game/Engine/Engine.context.ts
@@ -1,5 +1,0 @@
-import { createContext } from 'react'
-
-import { Context } from 'react'
-
-export const engineContext: Context<null> = createContext<null>(null)
