--- client/src/components/_Game/Engine/Engine.interface.ts
@@ -1,8 +1,0 @@
-import { RootState } from '@react-three/fiber'
-
-export interface EngineProps {
-  children: JSX.Element
-  getRenderer: () => RootState['gl']
-  getCamera: () => RootState['camera']
-  getScene: () => RootState['scene']
-}
