--- client/src/components/_Game/units/common/Hero/Hero.interface.ts
@@ -1,5 +1,0 @@
-import { GroupProps } from '@react-three/fiber'
-
-export interface HeroProps {
-  groupProps?: GroupProps
-}
