--- client/src/store/app/interface.ts
@@ -1,6 +1,0 @@
-import * as YUKA from 'yuka'
-
-export interface AppStore {
-  fps: number
-  entityManager: YUKA.EntityManager
-}
