--- client/src/store/player/interface.ts
@@ -1,9 +1,0 @@
-import { Hero } from '@/interfaces/hero'
-import { Unit } from '@/interfaces/unit'
-
-export interface PlayerStore {
-  unitId: Unit['id']
-  getHero: GetHero
-}
-
-export type GetHero = () => Hero | null
